requirements.txt
flask
rembg
Pillow
werkzeug
